+++ src/signals-factory.ts
@@ -2,20 +2,20 @@
 
 import { Store } from './store';
 
-interface SetupWithStore {
+type SetupWithStore = {
   readonly setup: (store: Store) => void;
-}
+};
 
-interface SignalsTypeWrapper<SignalsType> {
-  signals: SignalsType;
-}
+type SignalsTypeWrapper<SignalsType> = {
+  readonly signals: SignalsType;
+};
 
 export type Signals<SignalsType> = SetupWithStore & SignalsTypeWrapper<SignalsType>;
 
-export interface MappedSignalsType<SignalsType1, SignalsType2> {
-  readonly signals1: SignalsType1;
-  readonly signals2: SignalsType2;
-}
+export type MappedSignalsType<SignalsType1, SignalsType2> = Readonly<{
+  signals1: SignalsType1;
+  signals2: SignalsType2;
+}>;
 
 /**
  * This is the interface for signal factories, which represent a higher abstraction over the usage
@@ -24,21 +24,26 @@ export interface MappedSignalsType<SignalsType1, SignalsType2> {
  * Use the createSignalsFactory utility function to create your own SignalsFactory (see the implementation
  * of EffectSignalsFactory for an example).
  *
- * @typedef {object} SignalsFactory<SignalsType> - interface for monadic signal factories
- * @template SignalsType - specifies the interface for signals provided by the factory (type identifiers)
+ * @typedef {object} SignalsFactory<SignalsType> - type for monadic signal factories
+ * @template SignalsType - specifies the type for signals provided by the factory (type identifiers)
  * @property {function} build - returns an object with a setup function (taking a store as argument) and the signals being setup
  * @property {function} bind - the monadic bind (aka flatMap) to compose with other signal factories
- * @property {function} fmap - (aka map) a functor to map the signals produced by the factory
+ * @property {function} fmap - (aka map) the functor map, to map the signals produced by the factory
  */
-export interface SignalsFactory<SignalsType> {
-  readonly build: () => Signals<SignalsType>;
-  readonly bind: <SignalsType2>(
+export type SignalsFactory<SignalsType> = Readonly<{
+  build: () => Signals<SignalsType>;
+  bind: <SignalsType2>(
     mapper: (signals: Signals<SignalsType>) => SignalsFactory<SignalsType2>,
   ) => SignalsFactory<MappedSignalsType<SignalsType, SignalsType2>>;
-  readonly fmap: <SignalsType2>(
+  fmap: <SignalsType2>(
     mapper: (signals: Signals<SignalsType>) => Signals<SignalsType2>,
   ) => SignalsFactory<SignalsType2>;
-}
+}>;
+
+type SignalsFactoryMapCreate = <SignalsType1, SignalsType2>(
+  factory1: SignalsFactory<SignalsType1>,
+  mapper: (signals: Signals<SignalsType1>) => Signals<SignalsType2>,
+) => SignalsFactory<SignalsType2>;
 
 /**
  * A utility function that implements fmap for signal factories. However, instead of using this
@@ -50,7 +55,7 @@ export interface SignalsFactory<SignalsType> {
  * @param {function} mapper - the function mapping from Signals<SignalsType1> to Signals<SignalsType2>
  * @returns {SignalsFactory<SignalsType2>} the resulting factory
  */
-export const signalsFactoryMap = <SignalsType1, SignalsType2>(
+export const signalsFactoryMap: SignalsFactoryMapCreate = <SignalsType1, SignalsType2>(
   factory1: SignalsFactory<SignalsType1>,
   mapper: (signals: Signals<SignalsType1>) => Signals<SignalsType2>,
 ): SignalsFactory<SignalsType2> => {
@@ -70,6 +75,11 @@ export const signalsFactoryMap = <SignalsType1, SignalsType2>(
   return factory2;
 };
 
+type SignalsFactoryBindCreate = <SignalsType1, SignalsType2>(
+  factory1: SignalsFactory<SignalsType1>,
+  mapper: (signals: Signals<SignalsType1>) => SignalsFactory<SignalsType2>,
+) => SignalsFactory<MappedSignalsType<SignalsType1, SignalsType2>>;
+
 /**
  * A utility function that implements bind for signal factories. However, instead of using this
  * low-level function, in most cases you should just use the createSignalsFactory utility function.
@@ -80,7 +90,7 @@ export const signalsFactoryMap = <SignalsType1, SignalsType2>(
  * @param {function} mapper - the function mapping from Signals<SignalsType1> to SignalsFactory<SignalsType2>
  * @returns {SignalsFactory<SignalsType2>} the resulting factory
  */
-export const signalsFactoryBind = <SignalsType1, SignalsType2>(
+export const signalsFactoryBind: SignalsFactoryBindCreate = <SignalsType1, SignalsType2>(
   factory1: SignalsFactory<SignalsType1>,
   mapper: (signals: Signals<SignalsType1>) => SignalsFactory<SignalsType2>,
 ): SignalsFactory<MappedSignalsType<SignalsType1, SignalsType2>> => {
@@ -118,6 +128,10 @@ export const signalsFactoryBind = <SignalsType1, SignalsType2>(
   return factory2;
 };
 
+type SignalsFactoryCreate = <SignalsType>(
+  build: () => Signals<SignalsType>,
+) => SignalsFactory<SignalsType>;
+
 /**
  * This utility function creates an object that implements the SignalsFactory interface. It should be
  * used to implement all specific signal factories (by just extending the returned object as required).
@@ -127,7 +141,7 @@ export const signalsFactoryBind = <SignalsType1, SignalsType2>(
  * @param {function} build - the function that implements build for SignalsFactory<SignalsType>
  * @returns {SignalsFactory<SignalsType>}
  */
-export const createSignalsFactory = <SignalsType>(
+export const createSignalsFactory: SignalsFactoryCreate = <SignalsType>(
   build: () => Signals<SignalsType>,
 ): SignalsFactory<SignalsType> => {
   let factory: SignalsFactory<SignalsType>;