+++ test/basic-behaviors.test.ts
@@ -1,11 +1,11 @@
 import { interval, NEVER, of } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
 import { Store } from '../src/store';
-import { TypeIdentifier } from '../src/store.utils';
+import { getIdentifier } from '../src/store.utils';
 import { awaitCompletion, awaitError, expectSequence } from './test.utils';
 
 describe('Behavior basics', () => {
-  const testId: TypeIdentifier<number> = { symbol: Symbol('TestBehavior') };
+  const testId = getIdentifier<number>();
 
   let store: Store;
 
@@ -19,7 +19,7 @@ describe('Behavior basics', () => {
       expect(() => {
         store.addLazyBehavior(testId, NEVER);
       }).toThrowError(
-        'A behavior or event source with the given identifier was already added with a source: Symbol(TestBehavior)',
+        'A behavior or event source with the given identifier was already added with a source: Symbol()',
       );
     });
     it('should be possible to add behavior sources again, after initial source has completed', async () => {
@@ -58,7 +58,7 @@ describe('Behavior basics', () => {
       expect(() => {
         store.addNonLazyBehavior(testId, NEVER);
       }).toThrowError(
-        'A behavior or event source with the given identifier was already added with a source: Symbol(TestBehavior)',
+        'A behavior or event source with the given identifier was already added with a source: Symbol()',
       );
     });