+++ test/complex-stateful-query-effect.test.ts
@@ -1,6 +1,6 @@
 import { combineLatest, of } from 'rxjs';
 import { filter, map, switchMap, withLatestFrom } from 'rxjs/operators';
-import { TypeIdentifier } from '../src/store.utils';
+import { getIdentifier } from '../src/store.utils';
 import { Store } from './../src/store';
 import { awaitStringifyEqualState, expectSequence } from './test.utils';
 describe('Stateful query pattern', () => {
@@ -14,11 +14,11 @@ describe('Stateful query pattern', () => {
     resultQuery: QueryType | null;
   }
 
-  const queryBehavior: TypeIdentifier<QueryType> = { symbol: Symbol('QueryBehavior') };
-  const resultBehavior: TypeIdentifier<ResultType> = { symbol: Symbol('ResultBehavior') };
-  const loadingBehavior: TypeIdentifier<boolean> = { symbol: Symbol('LoadingBehavior') };
-  const queryEvent: TypeIdentifier<Partial<QueryType>> = { symbol: Symbol('QueryEvent') };
-  const resultEvent: TypeIdentifier<ResultType> = { symbol: Symbol('ResultEvent') };
+  const queryBehavior = getIdentifier<QueryType>();
+  const resultBehavior = getIdentifier<ResultType>();
+  const loadingBehavior = getIdentifier<boolean>();
+  const queryEvent = getIdentifier<Partial<QueryType>>();
+  const resultEvent = getIdentifier<ResultType>();
   const resultEffect = Symbol('ResultEffect');
 
   let store: Store;